Moved Lecturer Part to another form for work efficiency.
